DBZ-9160 refactoring integration tests to address the volume of features
