+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/dto/Order.java
@@ -3,8 +3,8 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.quarkus.sample.app;
 
-record EngineManifest(String group, String connector) {
+package io.quarkus.sample.app.dto;
 
+public record Order(int id, String name, String description) {
 }
