+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/dto/Product.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.sample.app;
+package io.quarkus.sample.app.dto;
 
-public record Order(int id, String name, String description) {
+public record Product(int id, String name, String description) {
 }
