+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/resources/CapturingResource.java
@@ -3,42 +3,31 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
+package io.quarkus.sample.app.resources;
 
-package io.quarkus.sample.app;
-
-import jakarta.inject.Inject;
+import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.core.Response;
 
-import io.debezium.runtime.DebeziumConnectorRegistry;
-import io.debezium.runtime.DebeziumStatus;
+import io.quarkus.sample.app.services.OrderService;
+import io.quarkus.sample.app.services.ProductService;
 
-@Path("/api/debezium")
-public class DebeziumEndpoint {
+@Path("/captured")
+@ApplicationScoped
+public class CapturingResource {
 
-    private final DebeziumConnectorRegistry registry;
     private final ProductService productService;
     private final OrderService orderService;
 
-    @Inject
-    public DebeziumEndpoint(DebeziumConnectorRegistry registry,
-                            ProductService productService,
-                            OrderService orderService) {
-        this.registry = registry;
+    public CapturingResource(ProductService productService, OrderService orderService) {
         this.productService = productService;
         this.orderService = orderService;
     }
 
     @GET
-    @Path("status")
-    public DebeziumStatus getState() {
-        return registry.engines().getFirst().status();
-    }
-
-    @GET
-    @Path("captured")
-    public Response capture() {
+    @Path("all")
+    public Response captured() {
         if (productService.isInvoked()) {
             return Response.status(Response.Status.FOUND).build();
         }
@@ -53,15 +42,6 @@ public class DebeziumEndpoint {
             return Response.status(Response.Status.NOT_FOUND).build();
         }
         return Response.ok(productService.products()).build();
-    }
-
-    @GET
-    @Path("engines")
-    public Response engines() {
-        return Response.ok(registry.engines()
-                .stream()
-                .map(engine -> new EngineManifest(engine.captureGroup().id(), engine.connector().name()))
-                .toList()).build();
     }
 
     @GET
