+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/resources/HeartbeatEventResource.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.sample.app.events;
+package io.quarkus.sample.app.resources;
 
 import jakarta.inject.Inject;
 import jakarta.ws.rs.GET;
@@ -13,6 +13,7 @@ import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 
 import io.debezium.runtime.events.DebeziumHeartbeat;
+import io.quarkus.sample.app.events.HeartbeatEventObserver;
 
 @Path("/heartbeat")
 public class HeartbeatEventResource {
