+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/services/OrderService.java
@@ -4,12 +4,14 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.sample.app;
+package io.quarkus.sample.app.services;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import jakarta.enterprise.context.ApplicationScoped;
+
+import io.quarkus.sample.app.dto.Order;
 
 @ApplicationScoped
 public class OrderService {
