+++ quarkus-debezium-parent/quarkus-debezium-postgres-parent/integration-test/src/main/java/io/quarkus/sample/app/services/ProductService.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.sample.app;
+package io.quarkus.sample.app.services;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,12 +12,14 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import jakarta.enterprise.context.ApplicationScoped;
 
+import io.quarkus.sample.app.dto.Product;
+
 @ApplicationScoped
 public class ProductService {
     private final List<Product> products = new ArrayList<>();
     private final AtomicBoolean invoked = new AtomicBoolean(false);
 
-    public void capture() {
+    public void captured() {
         invoked.set(true);
     }
